requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
pillow==10.2.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
